hook化
